requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
pdfplumber>=0.9.0
html2text>=2020.1.16
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
torch>=1.9.0
scikit-learn>=1.0.0
graphviz>=0.20.0
pytest>=7.4.0
